+++ templates/statistics-tpl.php
@@ -6,7 +6,7 @@ include dirname(__DIR__)."/lib/Auth/Process/DatabaseCommand.php";
 $this->data['header'] = 'SimpleSAMLphp Statistics';
 
 $this->data['jquery'] = array('core' => TRUE, 'ui' => TRUE, 'css' => TRUE);
-$this->data['head'] = '<link rel="stylesheet"  media="screen" type="text/css" href="' . SimpleSAML_Module::getModuleUrl('proxystatistics/statisticsproxy.css')  . '" />';
+$this->data['head'] = '<link rel="stylesheet"  media="screen" type="text/css" href="' . SimpleSAML\Module::getModuleUrl('proxystatistics/statisticsproxy.css')  . '" />';
 $this->data['head'] .='';
 $this->data['head'] .= '<script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>';
 $this->data['head'] .= '<script type="text/javascript">